refactor(api): move checkAuthorizationToManageCampaign prehandler to prescription/campaign
