+++ api/src/prescription/campaign/application/usecases/checkAuthorizationToManageCampaign.js
@@ -1,5 +1,5 @@
-import * as prescriberRoleRepository from '../../infrastructure/repositories/prescriber-role-repository.js';
-import { CampaignAuthorization } from '../pre-handlers/CampaignAuthorization.js';
+import { CampaignAuthorization } from '../../../../shared/application/pre-handlers/CampaignAuthorization.js';
+import * as prescriberRoleRepository from '../../../../shared/infrastructure/repositories/prescriber-role-repository.js';
 
 const execute = async function ({ userId, campaignId }) {
   const prescriberRole = await prescriberRoleRepository.getForCampaign({ userId, campaignId });
